Rename datatip-provider -> datatip
