+++ lib/services/consumed/index.js
@@ -2,6 +2,6 @@
 
 import statusBar from "./status-bar";
 import autocomplete from "./autocomplete";
-import datatipProvider from "./datatip-provider";
+import datatipProvider from "./datatip";
 
 export default { statusBar, autocomplete, datatipProvider };
